feat(buildings): Les Halles UsagePlan + MixedUseLesHallesMonotower
Extract commercial/livable gallery paint into `LesHallesUsagePlan` types
(regions-only). Add `sample_monotower` for a shared balanced shell, and
stack commercial-below / livable-above storeys with frozen shaft slots
(no stairs yet).

## Agent Dialogue
- **Developer:** Design MixedUseLesHallesMonotower; UsagePlan trait; shared shell; stairs later.
- **Agent:** Clarified plan vs regions (paint takes `FillableRegions` only). Implemented UsagePlan extraction, monotower sampling (XZ-stable noise), and the first vertical stack.
