+++ maybraid/richmond/buildings/src/monotower.rs
@@ -5,13 +5,12 @@
 //!
 //! The monotower has the responsibility of filling in proper shafts.
 //!
-//! When monotowers fit and AaBb3d, they choose a floor height from a range
-//! and then adjust to have a whole number of storeys.
+//! When monotowers fit an [`Aabb3d`](bevy_math::bounding::Aabb3d), they choose a
+//! floor height from a range and then adjust to have a whole number of storeys.
 //!
-//! Monotowers typically follow a Parameterized -> FloorPlan -> Full approach as well
-//! wherein floor plans include storey floor plans and stairs in the shafts. Then usage
-//! areas are filled using the same pattern as a particular full variant per storey.
+//! Monotowers typically follow a Parameterized → FloorPlan → UsagePlan approach:
+//! floor plans carry the shared shell (and shaft slots); usage plans paint
+//! residuals per storey (commercial / livable / …). Stairs in shafts are filled
+//! by the monotower once the stack is in place.
 //!
-//! Note that, going forward, we should probably make clearer the *UsagePlan concept in stories as well,
-//! breaking out the UsagePlan paint-on logic from the full-storey struct itself in a nice
-//! reusable way.
+//! Concrete towers live under [`crate::monotowers`].
