+++ maybraid/richmond/buildings/src/monotowers/mod.rs
@@ -1,0 +1,7 @@
+//! Monotowers: vertical stacks of a single floor-plan typology with continuous shafts.
+
+pub mod mixed_use_les_halles;
+
+pub use mixed_use_les_halles::{
+	MixedUseLesHallesMonotower, MixedUseLesHallesStorey,
+};
