+++ maybraid/richmond/buildings/src/storeys/les_halles/full_storey.rs
@@ -1,7 +1,7 @@
 //! Les Halles full storey: floor plan plus commercial gallery strip fills.
 //!
 //! The ring shell lives on [`LesHallesFloorPlan`]. This type fits that plan and
-//! fills [`SpaceKind::ExternalSpace`] strips with [`CommercialStallStrip`].
+//! paints [`LesHallesCommercialUsage`] onto [`SpaceKind::ExternalSpace`] strips.
 //! Residual walkways / shafts remain in [`FillableRegions::within`].
 
 use lod::gen::LodSceneLevel;
@@ -11,61 +11,33 @@ use richmond_building_components::labels::LabelNode;
 use richmond_building_components::panels::PanelNode;
 use richmond_building_components::{BuildingComponents, Layers};
 
-use crate::fit::{Confines, FillableRegions, Fit, FitError, SpaceKind};
+use crate::fit::{Confines, FillableRegions, Fit, FitError};
 use crate::usage_areas::CommercialStallStrip;
 
 use super::floor_plan::LesHallesFloorPlan;
+use super::usage_plan::{LesHallesCommercialUsage, LesHallesUsagePlan};
 
-/// Full Les Halles storey: shell on [`Self::floor_plan`] plus gallery stall strips.
+/// Full Les Halles storey: shell on [`Self::floor_plan`] plus commercial usage.
 #[derive(Debug, Clone, PartialEq)]
 pub struct LesHallesFullStorey {
 	pub floor_plan: LesHallesFloorPlan,
-	pub stall_strips: Vec<CommercialStallStrip>,
+	pub usage: LesHallesCommercialUsage,
 }
 
 impl LesHallesFullStorey {
-	/// Wrap an already-fitted floor plan and fill external gallery strips.
+	/// Gallery stall strips (same as [`Self::usage`].stall_strips).
+	pub fn stall_strips(&self) -> &[CommercialStallStrip] {
+		&self.usage.stall_strips
+	}
+
+	/// Wrap an already-fitted floor plan and paint commercial usage.
 	pub fn from_floor_plan(
 		floor_plan: LesHallesFloorPlan,
 		noise: NoiseParams,
 	) -> Result<(Self, FillableRegions), FitError> {
 		let regions = floor_plan.fillable_regions();
-		Self::fill_from_regions(floor_plan, regions, noise)
-	}
-
-	fn fill_from_regions(
-		floor_plan: LesHallesFloorPlan,
-		regions: FillableRegions,
-		noise: NoiseParams,
-	) -> Result<(Self, FillableRegions), FitError> {
-		let mut stall_strips = Vec::new();
-		let mut residual_within = Vec::new();
-		for (i, region) in regions.within.into_iter().enumerate() {
-			if region.kind != SpaceKind::ExternalSpace {
-				residual_within.push(region);
-				continue;
-			}
-			let mut strip_noise = noise;
-			strip_noise.seed = noise.seed.wrapping_add(i as i32 * 31);
-			match CommercialStallStrip::fit_to_confines(&region.confines, strip_noise) {
-				Ok((strip, _)) => stall_strips.push(strip),
-				Err(FitError::TooSmall { .. }) => {
-					// Leave unfilled if the strip is too narrow after shaft clears.
-					residual_within.push(region);
-				}
-				Err(err) => return Err(err),
-			}
-		}
-		Ok((
-			Self {
-				floor_plan,
-				stall_strips,
-			},
-			FillableRegions {
-				within: residual_within,
-				atop: regions.atop,
-			},
-		))
+		let (usage, residual) = LesHallesCommercialUsage::paint(regions, noise)?;
+		Ok((Self { floor_plan, usage }, residual))
 	}
 }
 
@@ -75,16 +47,15 @@ impl Fit for LesHallesFullStorey {
 		noise: NoiseParams,
 	) -> Result<(Self, FillableRegions), FitError> {
 		let (floor_plan, regions) = LesHallesFloorPlan::fit_to_confines(confines, noise)?;
-		Self::fill_from_regions(floor_plan, regions, noise)
+		let (usage, residual) = LesHallesCommercialUsage::paint(regions, noise)?;
+		Ok((Self { floor_plan, usage }, residual))
 	}
 }
 
 impl BuildingComponents for LesHallesFullStorey {
 	fn panel_nodes_for_level(&self, level: LodSceneLevel) -> Layers<PanelNode> {
 		let mut out = self.floor_plan.panel_nodes_for_level(level);
-		for strip in &self.stall_strips {
-			out.extend(strip.panel_nodes_for_level(level));
-		}
+		out.extend(self.usage.panel_nodes_for_level(level));
 		out
 	}
 
@@ -93,11 +64,7 @@ impl BuildingComponents for LesHallesFullStorey {
 	}
 
 	fn label_nodes_for_level(&self, level: LodSceneLevel) -> Layers<LabelNode> {
-		let mut out = Layers::new();
-		for strip in &self.stall_strips {
-			out.extend(strip.label_nodes_for_level(level));
-		}
-		out
+		self.usage.label_nodes_for_level(level)
 	}
 }
 
@@ -110,10 +77,11 @@ mod tests {
 	use procedural_common::NoiseParams;
 	use richmond_building_components::BuildingComponents;
 
+	use crate::fit::SpaceKind;
+	use crate::storeys::les_halles::{LesHallesFloorPlan, LesHallesParameterized};
+
 	#[test]
 	fn full_storey_fills_external_strips_with_labels() {
-		use crate::storeys::les_halles::{LesHallesFloorPlan, LesHallesParameterized};
-
 		let bounds = Aabb3d::from_min_max(
 			Vec3::new(-24.0, 0.0, -18.0),
 			Vec3::new(24.0, 4.0, 18.0),
@@ -125,7 +93,7 @@ mod tests {
 		let confines = Confines::new(bounds, 0.0, openings);
 		let (plan, _) = LesHallesFloorPlan::from_parameterized(params, &confines).unwrap();
 		let (storey, regions) = LesHallesFullStorey::from_floor_plan(plan, noise).unwrap();
-		assert!(!storey.stall_strips.is_empty());
+		assert!(!storey.stall_strips().is_empty());
 		assert!(regions
 			.within
 			.iter()
